requirements.txt
streamlit==1.54.0
azure-ai-projects==1.0.0
azure-identity